Replace default js confirm dialog with a custom service based dialog

+++ webclient/src/app/messages/messages.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component({
@@ -15,7 +16,7 @@ export class MessagesComponent implements OnInit {
   pageNumber = 1;
   pageSize = 10;
 
-  constructor(private messageService: MessageService) {}
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) {}
 
   ngOnInit(): void {
     this.loadMessages();
@@ -30,12 +31,16 @@ export class MessagesComponent implements OnInit {
   }
 
   deleteMessage(id: number) {
-    this.messageService.deleteMessage(id).subscribe(() =>
-      this.messages.splice(
-        this.messages.findIndex((m) => m.id === id),
-        1
-      )
-    );
+    this.confirmService.confirm('Confirm Delete Message', 'This cannot be undone').subscribe((result) => {
+      if (result) {
+        this.messageService.deleteMessage(id).subscribe(() =>
+          this.messages.splice(
+            this.messages.findIndex((m) => m.id === id),
+            1
+          )
+        );
+      }
+    });
   }
 
   pageChanged(event: any) {
